Added functions for displaying beziar curves
Made a beziar curve class to handle the creation and functioning of beziar curves. 

Includes a function to split beziar curves into shorter curves on the same path. the math is cool but complected. not quite sure why it works. something about derivatives. I've tested the theory but not my implementation.

Includes a function to make a line of svg code that matches the curve. 
Tested it. seems to work.