Got the split function working!
The beziar curve was being plotted with relative positions, but the math was being done with absolute positions. did not work.

added classes for grouping curves together into a frame, and for grouping frames into animation. plan to make order and content of frames programmatic.

improved debug a bit.

that's about it