added ability to change positions
Moved creating properly formatted svg strings into a separate function.
Added a function to set the position of each element in an svg string relative to the origin. For some reason it seems to remove some elements.

I now have access to the raw numbers that make up the vectors!