Improved Bezair Methods (again)
Fixed the Beziar methods to use the Inkscape format:
Multipule curves that attach to each other, use the last point of one as the first point of the next, meaning the next curve is relative to a point which is relative (and so on) to the origin. The number of points is no longer fixed at 4, but will always be 3x+1 where x is the number of concatenated curves.